#85ztnqybg Fix design and geo permission
