+++ app/src/main/res/layout/fragment_home.xml
@@ -126,7 +126,6 @@
                         <LinearLayout
                             android:layout_width="wrap_content"
                             android:layout_height="match_parent"
-                            android:layout_marginEnd="16dp"
                             android:layout_marginBottom="16dp"
                             android:orientation="horizontal">
